+++ docs/changelog.md
@@ -124,7 +124,7 @@ description: |
 - Miscellaneous Improvements
   - Added [`bookmarked`]($heading.bookmarked) argument to heading to control
     whether a heading becomes part of the PDF outline
-  - Added [`caption-pos`]($figure.caption-pos) argument to control the position
+  - Added [`caption-pos`]($figure.caption.pos) argument to control the position
     of a figure's caption
   - Added [`metadata`]($metadata) function for exposing an arbitrary value to
     the introspection system